Reduce MCP tools to a single list_roster for integration testing
Drop the payroll-specific tools (employee detail, PTOs, vacation balance,
reimbursements, payroll summary) and keep only a no-arg list_roster tool
that returns active employees with name, role, project, client and country.
The auth, security chain, JWT converter and protected-resource metadata are
unchanged.
