+++ src/main/java/com/entropyteam/entropay/mcp/dtos/RosterEntry.java
@@ -2,10 +2,6 @@ package com.entropyteam.entropay.mcp.dtos;
 
 import java.util.UUID;
 
-/**
- * Slim payroll roster row exposed by the MCP {@code list_payroll_roster} tool. Carries no
- * salary or personal data — those require {@code get_employee}.
- */
 public record RosterEntry(
         UUID id,
         String internalId,
